+++ app/src/main/java/eu/kanade/presentation/more/settings/PreferenceItem.kt
@@ -156,6 +156,7 @@ internal fun PreferenceItem(
                     subtitle = item.subtitle,
                     icon = item.icon,
                     value = values,
+                    isSecret = item.isSecret,
                     onConfirm = {
                         val accepted = item.onValueChanged(it)
                         if (accepted) item.preference.set(it)